requirements.txt
flask
mysql-connector-python
python-dotenv
tensorflow
numpy
opencv-python
scikit-learn
requests
pillow
pymongo
boto3
requests-aws4auth
opensearch-dsl
numpy<2



